+++ Fyber/src/main/java/com/mopub/mobileads/FyberAdapterConfiguration.java
@@ -29,11 +29,11 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM_WITH_THRO
 public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
 
     public final static String KEY_FYBER_APP_ID = "appID";
-    public final static String KEY_FYBER_DEBUG = "debug";
 
-    private static final String TAG = "FyberAdapterConfig";
-    private static final String MOPUB_NETWORK_NAME = BuildConfig.NETWORK_NAME;
+    private static final String ADAPTER_NAME = FyberAdapterConfiguration.class.getSimpleName();
     private static final String ADAPTER_VERSION = BuildConfig.VERSION_NAME;
+    private final static String KEY_FYBER_DEBUG = "debug";
+    private static final String MOPUB_NETWORK_NAME = BuildConfig.NETWORK_NAME;
 
     @NonNull
     @Override
@@ -65,8 +65,16 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
         Preconditions.checkNotNull(context);
         Preconditions.checkNotNull(listener);
 
+        if (configuration == null || configuration.isEmpty()) {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Fyber initialization failed. Configuration is null." +
+                    "Note that initialization on the first app launch is a no-op. It will be attempted again on subsequent ad requests.");
+            listener.onNetworkInitializationFinished(FyberAdapterConfiguration.class, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            return;
+        }
+
         if (configuration != null && !configuration.isEmpty()) {
             final String appId = configuration.get(KEY_FYBER_APP_ID);
+
             if (!TextUtils.isEmpty(appId)) {
                 initializeFyberMarketplace(context, appId,
                                            configuration.containsKey(KEY_FYBER_DEBUG),
@@ -75,13 +83,11 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
                                                public void onFyberAdapterConfigurationResolved(
                                                        OnFyberMarketplaceInitializedListener.FyberInitStatus status) {
 
-                                                   if (status ==
-                                                       OnFyberMarketplaceInitializedListener.FyberInitStatus.SUCCESSFULLY ||
-                                                           status ==
-                                                                   OnFyberMarketplaceInitializedListener.FyberInitStatus.FAILED) {
+                                                   if (status == OnFyberMarketplaceInitializedListener.FyberInitStatus.SUCCESSFULLY ||
+                                                           status == OnFyberMarketplaceInitializedListener.FyberInitStatus.FAILED) {
                                                        listener.onNetworkInitializationFinished(FyberAdapterConfiguration.class, MoPubErrorCode.ADAPTER_INITIALIZATION_SUCCESS);
                                                    } else if (status == OnFyberMarketplaceInitializedListener.FyberInitStatus.INVALID_APP_ID) {
-                                                       log("Attempted to initialize Fyber MarketPlace with wrong app id - " + appId);
+                                                       MoPubLog.log(CUSTOM, "Attempted to initialize Fyber Marketplace with wrong app id - " + appId);
                                                        listener.onNetworkInitializationFinished(FyberAdapterConfiguration.class, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
                                                    } else {
                                                        listener.onNetworkInitializationFinished(FyberAdapterConfiguration.class, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
@@ -89,7 +95,9 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
                                                }
                                            });
             } else {
-                MoPubLog.log(CUSTOM_WITH_THROWABLE, "No Fyber app id given in configuration object. Initialization postponed. You can use FyberAdapterConfiguration.KEY_FYBER_APP_ID as your configuration key");
+                MoPubLog.log(CUSTOM_WITH_THROWABLE, "No Fyber app id given in configuration object. " +
+                        "Initialization postponed until the next ad request.");
+                listener.onNetworkInitializationFinished(FyberAdapterConfiguration.class, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
 
             }
         }
@@ -98,7 +106,6 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
     public static void initializeFyberMarketplace(Context context, String appId, boolean debugMode, @NonNull
     final OnFyberAdapterConfigurationResolvedListener listener) {
         synchronized (FyberAdapterConfiguration.class) {
-            // Just to be on the safe side, wrap initialize with exception handling
             if (debugMode) {
                 InneractiveAdManager.setLogLevel(Log.VERBOSE);
             }
@@ -107,14 +114,12 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
                 InneractiveAdManager.initialize(context, appId,
                                                 new OnFyberMarketplaceInitializedListener() {
                                                     @Override
-                                                    public void onFyberMarketplaceInitialized(
-                                                            FyberInitStatus status) {
+                                                    public void onFyberMarketplaceInitialized(FyberInitStatus status) {
                                                         listener.onFyberAdapterConfigurationResolved(status);
-                                                    }
-                                                });
+                                                    }});
             } else if (!appId.equals(InneractiveAdManager.getAppId())) {
                 MoPubLog.log(CUSTOM, "Fyber Marketplace was initialized with appId " + InneractiveAdManager.getAppId() +
-                        " and now requests initialization with another appId (" + appId + ") You may have configured the wrong appId on the MoPub console?\n" +
+                        " and now requests initialization with another appId (" + appId + ") You may have configured the wrong appId on the MoPub console.\n" +
                         " you can only use a single appId and it's related spots");
                 listener.onFyberAdapterConfigurationResolved(
                         OnFyberMarketplaceInitializedListener.FyberInitStatus.INVALID_APP_ID);
@@ -126,16 +131,13 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
         }
     }
 
-    public static void updateGdprConsentStatusFromMopub() {
-        InneractiveAdManager.GdprConsentSource gdprConsentSource = InneractiveAdManager.getGdprStatusSource();
+    public static void updateGdprConsentStatus() {
+        final Boolean mopubGdpr = extractGdprFromMoPub();
 
-        if (gdprConsentSource == null || gdprConsentSource == InneractiveAdManager.GdprConsentSource.External) {
-            final Boolean mopubGdpr = extractGdprFromMoPub();
-            if (mopubGdpr == null) {
-                InneractiveAdManager.clearGdprConsentData();
-            } else {
-                InneractiveAdManager.setGdprConsent(mopubGdpr, InneractiveAdManager.GdprConsentSource.External);
-            }
+        if (mopubGdpr == null) {
+            InneractiveAdManager.clearGdprConsentData();
+        } else {
+            InneractiveAdManager.setGdprConsent(mopubGdpr, InneractiveAdManager.GdprConsentSource.External);
         }
     }
 
@@ -145,14 +147,14 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
         if (personalInfoManager != null && personalInfoManager.gdprApplies() == Boolean.TRUE) {
             /* Only set the GDPR consent flag, if GDPR is applied. If GDPR is not applied,
             canCollectPersonalInformation returns true, but there is no explicit consent*/
-                MoPubLog.log(CUSTOM, "Fyber sdk will user gdpr consent from mopub. GdprConsent- "
+                MoPubLog.log(CUSTOM, "Fyber will user GDPR consent collected by MoPub."
                         + personalInfoManager.canCollectPersonalInformation());
                 return personalInfoManager.canCollectPersonalInformation();
             } else if (personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.UNKNOWN && MoPub.shouldAllowLegitimateInterest()) {
-                MoPubLog.log(CUSTOM, "Gdpr result from mopub is unkown and publisher allowed liegitmateInterset. GdprConsent - true");
+                MoPubLog.log(CUSTOM, "GDPR result from MoPub is unknown and publisher allowed legitimate interest.");
                 return true;
             } else {
-                MoPubLog.log(CUSTOM, "Fyber sdk has not found any Gdpr values");
+                MoPubLog.log(CUSTOM, "Fyber has not found any GDPR values");
             }
         return null;
     }
@@ -161,7 +163,7 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
      * Internal interface to bridge the gap between the custom event classes and the initializeNetwork
      */
     public interface OnFyberAdapterConfigurationResolvedListener {
-        public void onFyberAdapterConfigurationResolved(
+        public void onFyberAdapterConfigurationResolved (
                 OnFyberMarketplaceInitializedListener.FyberInitStatus status);
     }
 
@@ -179,7 +181,7 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
                 try {
                     age = Integer.valueOf(extras.get(FyberMoPubMediationDefs.KEY_AGE));
                 } catch (NumberFormatException e) {
-                    log("local extra contains Invalid Age");
+                    MoPubLog.log(CUSTOM, "localExtras contains Invalid Age");
                 }
             }
 
@@ -216,9 +218,5 @@ public class FyberAdapterConfiguration extends BaseAdapterConfiguration {
                 request.setKeywords(keywords);
             }
         }
-    }
-
-    private static void log(String message) {
-        MoPubLog.log(CUSTOM, TAG, message);
     }
 }