+++ Fyber/src/main/java/com/mopub/mobileads/FyberBanner.java
@@ -43,8 +43,6 @@ import static com.mopub.mobileads.MoPubErrorCode.INLINE_LOAD_ERROR;
 public class FyberBanner extends BaseAd {
 
   private static final String ADAPTER_NAME = FyberBanner.class.getSimpleName();
-  private final static String LOG_TAG = "FyberBannerForMopub";
-
   private final FyberAdapterConfiguration mFyberAdapterConfiguration;
 
   public FyberBanner() {
@@ -55,14 +53,14 @@ public class FyberBanner extends BaseAd {
   private InneractiveAdSpot mBannerSpot;
   private ViewGroup mAdLayout;
 
-
   private void requestBanner(final Context context, String spotId, Map<String, String> localExtras) {
-    FyberAdapterConfiguration.updateGdprConsentStatusFromMopub();
+    FyberAdapterConfiguration.updateGdprConsentStatus();
 
     mSpotId = spotId;
 
     if (mBannerSpot != null) {
       mBannerSpot.destroy();
+      mBannerSpot = null;
     }
 
     mBannerSpot = InneractiveAdSpotManager.get().createSpot();
@@ -72,14 +70,17 @@ public class FyberBanner extends BaseAd {
     InneractiveAdViewUnitController controller = new InneractiveAdViewUnitController();
     mBannerSpot.addUnitController(controller);
   
-    InneractiveAdRequest request = new InneractiveAdRequest(spotId);
+    final InneractiveAdRequest request = new InneractiveAdRequest(spotId);
     FyberAdapterConfiguration.updateRequestFromExtras(request, localExtras);
 
     mBannerSpot.setRequestListener(new InneractiveAdSpot.RequestListener() {
       @Override
       public void onInneractiveSuccessfulAdRequest(InneractiveAdSpot adSpot) {
         if (adSpot != mBannerSpot) {
-          MoPubLog.log(LOAD_FAILED, "Wrong Banner Spot received - " + adSpot + ", Actual - " + mBannerSpot);
+          MoPubLog.log(getAdNetworkId(), LOAD_FAILED, "Wrong Banner Spot received - " + adSpot + ", Actual - " + mBannerSpot);
+          if (mLoadListener != null) {
+            mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+          }
           return;
         }
 
@@ -108,12 +109,12 @@ public class FyberBanner extends BaseAd {
         
           @Override
           public void onAdWillCloseInternalBrowser(InneractiveAdSpot adSpot) {
-            log("onAdWillCloseInternalBrowser");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdWillCloseInternalBrowser");
           }
         
           @Override
           public void onAdWillOpenExternalApp(InneractiveAdSpot adSpot) {
-            log("onAdWillOpenExternalApp");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdWillOpenExternalApp");
             // Don't call the onLeaveApplication() API since it causes a false Click event on MoPub
           }
         
@@ -123,18 +124,19 @@ public class FyberBanner extends BaseAd {
 
             if (error instanceof WebViewRendererProcessHasGoneError) {
               if (mInteractionListener != null) {
-                mInteractionListener.onAdFailed(MoPubErrorCode.RENDER_PROCESS_GONE_UNSPECIFIED);
                 MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME);
+                mInteractionListener.onAdFailed(MoPubErrorCode.RENDER_PROCESS_GONE_UNSPECIFIED);
               } else if (mLoadListener != null) {
-                mLoadListener.onAdLoadFailed(MoPubErrorCode.RENDER_PROCESS_GONE_UNSPECIFIED);
                 MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME);
+                mLoadListener.onAdLoadFailed(MoPubErrorCode.RENDER_PROCESS_GONE_UNSPECIFIED);
               }
             }
           }
         
           @Override
           public void onAdExpanded(InneractiveAdSpot adSpot) {
-            log("onAdExpanded");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdExpanded");
+
             if (mInteractionListener != null) {
               mInteractionListener.onAdExpanded();
             }
@@ -142,12 +144,13 @@ public class FyberBanner extends BaseAd {
         
           @Override
           public void onAdResized(InneractiveAdSpot adSpot) {
-            log("onAdResized");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdResized");
           }
         
           @Override
           public void onAdCollapsed(InneractiveAdSpot adSpot) {
-            log("onAdCollapsed");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdCollapsed");
+
             if (mInteractionListener != null) {
               mInteractionListener.onAdCollapsed();
             }
@@ -163,6 +166,9 @@ public class FyberBanner extends BaseAd {
       @Override
       public void onInneractiveFailedAdRequest(InneractiveAdSpot adSpot,
                                                InneractiveErrorCode errorCode) {
+        MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, INLINE_LOAD_ERROR.getIntCode(),
+                errorCode);
+
         if (mLoadListener != null) {
           if (errorCode == InneractiveErrorCode.CONNECTION_ERROR) {
             mLoadListener.onAdLoadFailed(MoPubErrorCode.NO_CONNECTION);
@@ -173,8 +179,6 @@ public class FyberBanner extends BaseAd {
           } else {
             mLoadListener.onAdLoadFailed(MoPubErrorCode.SERVER_ERROR);
           }
-          MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, INLINE_LOAD_ERROR.getIntCode(),
-                  errorCode);
         }
       }
     });
@@ -200,7 +204,7 @@ public class FyberBanner extends BaseAd {
   @NonNull
   @Override
   protected String getAdNetworkId() {
-    return mSpotId;
+    return mSpotId == null ? "" : mSpotId;
   }
 
   @Override
@@ -227,9 +231,9 @@ public class FyberBanner extends BaseAd {
     }
 
     final boolean isBannerFormat = "banner".equalsIgnoreCase(adUnitFormat);
-    final boolean isMrecFormat = "medium_rectangle".equalsIgnoreCase(adUnitFormat);
+    final boolean isMrectFormat = "medium_rectangle".equalsIgnoreCase(adUnitFormat);
 
-    if (!isBannerFormat && !isMrecFormat) {
+    if (!isBannerFormat && !isMrectFormat) {
       MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
               "Fyber only supports 320*50, 728*90 and 300*250 sized ads. " +
                       "Please ensure your MoPub ad unit's format is either Banner or Medium Rectangle.");
@@ -261,7 +265,7 @@ public class FyberBanner extends BaseAd {
                 @Override
                 public void onFyberAdapterConfigurationResolved(
                         OnFyberMarketplaceInitializedListener.FyberInitStatus status) {
-                  //note - we try to load ads when "FAILED" because an ad request will re-attempt to initialize the relevant parts of the SDK.
+                  //note - Fyber tries to load ads when "FAILED" because an ad request will re-attempt to initialize the relevant parts of the SDK.
                   if (status == OnFyberMarketplaceInitializedListener.FyberInitStatus.SUCCESSFULLY
                           || status == OnFyberMarketplaceInitializedListener.FyberInitStatus.FAILED) {
                     requestBanner(context, spotId, extras);
@@ -288,7 +292,7 @@ public class FyberBanner extends BaseAd {
    */
   protected void show() {
     if (mBannerSpot != null && mBannerSpot.getSelectedUnitController() != null && mAdLayout != null) {
-      InneractiveAdViewUnitController controller = (InneractiveAdViewUnitController) mBannerSpot
+      final InneractiveAdViewUnitController controller = (InneractiveAdViewUnitController) mBannerSpot
               .getSelectedUnitController();
 
       controller.bindView(mAdLayout);
@@ -306,11 +310,4 @@ public class FyberBanner extends BaseAd {
     return mAdLayout;
   }
 
-  /**
-   * MopubLog helper
-   * @param message
-   */
-  private void log(String message) {
-    MoPubLog.log(CUSTOM, LOG_TAG, message);
-  }
 }