+++ Fyber/src/main/java/com/mopub/mobileads/FyberRewardedVideo.java
@@ -35,15 +35,15 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOULD_REWARD;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
 import static com.mopub.mobileads.MoPubErrorCode.FULLSCREEN_LOAD_ERROR;
 
 public class FyberRewardedVideo extends BaseAd {
-    private final static String LOG_TAG = "FyberRewardedVideoForMopub";
-    private static final String ADAPTER_NAME = FyberRewardedVideo.class.getSimpleName();
 
+    private static final String ADAPTER_NAME = FyberRewardedVideo.class.getSimpleName();
     private final FyberAdapterConfiguration mFyberAdapterConfiguration;
 
     public FyberRewardedVideo() {
@@ -54,7 +54,6 @@ public class FyberRewardedVideo extends BaseAd {
 
     InneractiveAdSpot mRewardedSpot;
     Activity mParentActivity;
-    private boolean mRewarded = false;
 
     @Override
     protected LifecycleListener getLifecycleListener() {
@@ -63,11 +62,14 @@ public class FyberRewardedVideo extends BaseAd {
 
     @Override
     protected String getAdNetworkId() {
-        return mSpotId;
+        return mSpotId == null ? "" : mSpotId;
     }
 
     @Override
     protected boolean checkAndInitializeSdk(@NonNull Activity launcherActivity, @NonNull AdData adData) throws Exception {
+        Preconditions.checkNotNull(launcherActivity);
+        Preconditions.checkNotNull(adData);
+
         mParentActivity = launcherActivity;
         return false;
     }
@@ -75,6 +77,7 @@ public class FyberRewardedVideo extends BaseAd {
     @Override
     protected void load(@NonNull Context context, @NonNull AdData adData) throws Exception {
         Preconditions.checkNotNull(adData);
+        Preconditions.checkNotNull(context);
 
         MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
 
@@ -101,7 +104,7 @@ public class FyberRewardedVideo extends BaseAd {
                         @Override
                         public void onFyberAdapterConfigurationResolved(
                                 OnFyberMarketplaceInitializedListener.FyberInitStatus status) {
-                            //note - we try to load ads when "FAILED" because an ad request will re-attempt to initialize the relevant parts of the SDK.
+                            //note - Fyber tries to load ads when "FAILED" because an ad request will re-attempt to initialize the relevant parts of the SDK.
                             if (status == OnFyberMarketplaceInitializedListener.FyberInitStatus.SUCCESSFULLY || status == OnFyberMarketplaceInitializedListener.FyberInitStatus.FAILED) {
                                 requestRewarded(extras);
                             } else if (mLoadListener != null) {
@@ -136,23 +139,17 @@ public class FyberRewardedVideo extends BaseAd {
         MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
 
         if (mRewardedSpot != null && mRewardedSpot.isReady()) {
-
-            InneractiveFullscreenUnitController fullscreenUnitController = (InneractiveFullscreenUnitController)mRewardedSpot.getSelectedUnitController();
+            final InneractiveFullscreenUnitController fullscreenUnitController = (InneractiveFullscreenUnitController)mRewardedSpot.getSelectedUnitController();
             fullscreenUnitController.setEventsListener(new InneractiveFullscreenAdEventsListener() {
-
                 @Override
                 public void onAdDismissed(InneractiveAdSpot adSpot) {
-                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Fyber interstitial is dismissed");
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Fyber rewarded video has been dismissed");
 
                     if (mInteractionListener != null) {
                         mInteractionListener.onAdDismissed();
                     }
                 }
 
-                /**
-                 * Called by Fyber Marketplace when an interstitial ad activity is shown
-                 * @param adSpot Spot object
-                 */
                 @Override
                 public void onAdImpression(InneractiveAdSpot adSpot) {
                     MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
@@ -174,18 +171,19 @@ public class FyberRewardedVideo extends BaseAd {
 
                 @Override
                 public void onAdWillOpenExternalApp(InneractiveAdSpot adSpot) {
-                    log("onAdWillOpenExternalApp");
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdWillOpenExternalApp");
                     // Don't call the onLeaveApplication() API since it causes a false Click event on MoPub
                 }
 
                 @Override
                 public void onAdEnteredErrorState(InneractiveAdSpot adSpot, AdDisplayError error) {
-                    log("onAdEnteredErrorState - " + error.getMessage());
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdEnteredErrorState - " +
+                            error.getMessage());
                 }
 
                 @Override
                 public void onAdWillCloseInternalBrowser(InneractiveAdSpot adSpot) {
-                    log("onAdWillCloseInternalBrowser");
+                    MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdWillCloseInternalBrowser");
                 }
             });
 
@@ -199,6 +197,7 @@ public class FyberRewardedVideo extends BaseAd {
                 @Override
                 public void onCompleted() {
                     /* Got video content completed event. Do not report reward back just yet.*/
+                    MoPubLog.log(CUSTOM, ADAPTER_NAME, "Video content completed event. Do not report reward back yet.");
                 }
 
                 @Override
@@ -206,7 +205,7 @@ public class FyberRewardedVideo extends BaseAd {
                     MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME, "Video content play error event");
 
                     if (mInteractionListener != null) {
-                        mInteractionListener.onAdFailed(MoPubErrorCode.VIDEO_PLAYBACK_ERROR);
+                        mInteractionListener.onAdFailed(MoPubErrorCode.AD_SHOW_ERROR);
                     }
                 }
             });
@@ -215,9 +214,10 @@ public class FyberRewardedVideo extends BaseAd {
                 @Override
                 public void onAdRewarded(InneractiveAdSpot adSpot) {
                     if (mInteractionListener != null) {
-                        mInteractionListener.onAdComplete(mRewarded ?
-                                MoPubReward.success(MoPubReward.NO_REWARD_LABEL,
-                                        MoPubReward.DEFAULT_REWARD_AMOUNT) : MoPubReward.failure());
+                        MoPubLog.log(getAdNetworkId(), SHOULD_REWARD, ADAPTER_NAME, MoPubReward.DEFAULT_REWARD_AMOUNT,
+                                MoPubReward.NO_REWARD_LABEL);
+                        mInteractionListener.onAdComplete(MoPubReward.success(MoPubReward.NO_REWARD_LABEL,
+                                MoPubReward.DEFAULT_REWARD_AMOUNT));
                     }
                 }
             });
@@ -235,10 +235,11 @@ public class FyberRewardedVideo extends BaseAd {
     private void requestRewarded(Map<String, String> localExtras) {
 
         if (mParentActivity == null || TextUtils.isEmpty(mSpotId)) {
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
             mLoadListener.onAdLoadFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
         }
 
-        FyberAdapterConfiguration.updateGdprConsentStatusFromMopub();
+        FyberAdapterConfiguration.updateGdprConsentStatus();
 
         if (mRewardedSpot != null) {
             mRewardedSpot.destroy();
@@ -252,7 +253,7 @@ public class FyberRewardedVideo extends BaseAd {
         InneractiveFullscreenUnitController fullscreenUnitController = new InneractiveFullscreenUnitController();
         mRewardedSpot.addUnitController(fullscreenUnitController);
 
-        InneractiveAdRequest request = new InneractiveAdRequest(mSpotId);
+        final InneractiveAdRequest request = new InneractiveAdRequest(mSpotId);
         FyberAdapterConfiguration.updateRequestFromExtras(request, localExtras);
 
         mRewardedSpot.setRequestListener(new InneractiveAdSpot.RequestListener() {
@@ -268,26 +269,22 @@ public class FyberRewardedVideo extends BaseAd {
 
             @Override
             public void onInneractiveFailedAdRequest(InneractiveAdSpot adSpot, InneractiveErrorCode errorCode) {
+                MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, FULLSCREEN_LOAD_ERROR.getIntCode(),
+                        errorCode);
                 if (mLoadListener != null) {
                     if (errorCode == InneractiveErrorCode.CONNECTION_ERROR) {
                         mLoadListener.onAdLoadFailed(MoPubErrorCode.NO_CONNECTION);
-                    } else if  (errorCode == InneractiveErrorCode.CONNECTION_TIMEOUT) {
+                    } else if (errorCode == InneractiveErrorCode.CONNECTION_TIMEOUT) {
                         mLoadListener.onAdLoadFailed(MoPubErrorCode.NETWORK_TIMEOUT);
                     } else if (errorCode == InneractiveErrorCode.NO_FILL) {
-                        mLoadListener.onAdLoadFailed(MoPubErrorCode.NO_FILL);
+                        mLoadListener.onAdLoadFailed(MoPubErrorCode.NETWORK_NO_FILL);
                     } else {
-                        mLoadListener.onAdLoadFailed(MoPubErrorCode.SERVER_ERROR);
+                        mLoadListener.onAdLoadFailed(MoPubErrorCode.UNSPECIFIED);
                     }
                 }
-                MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, FULLSCREEN_LOAD_ERROR.getIntCode(),
-                        errorCode);
             }
         });
 
         mRewardedSpot.requestAd(request);
-    }
-
-    private void log(String message) {
-        MoPubLog.log(CUSTOM, LOG_TAG, message);
     }
 }