+++ Fyber/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Changelog
-  * 7.8.1.0
-    * Initial commit
+  * 7.8.2.0
     * Add support for banner, interstitial and rewarded ads.
-    * This version of the adapters has been certified with Fyber Marketplace 7.8.1 and MoPub 5.16.2.
+    * This version of the adapters has been certified with Fyber Marketplace 7.8.2 and MoPub 5.17.0.
+    * This and newer adapter versions are only compatible with 5.16.4 MoPub SDK